Create module capitalizeable which is used by all POROs
