+++ app/poros/background_image.rb
@@ -1,4 +1,5 @@
 class BackgroundImage
+  include Capitalizeable
   attr_reader :id, :place, :image_url, :source, :author
 
   def initialize(data, location)
@@ -12,12 +13,5 @@ class BackgroundImage
 
   def get_source(data)
     return 'pixabay.com' if data[:pageURL].include?('pixabay')
-  end
-
-  def capitalize_location(location)
-    split_location = location.split(',')
-    split_location[0].capitalize!
-    split_location[1].upcase!
-    split_location.join(', ')
   end
 end
